1581 clean up double-injections
