+++ core/src/main/java/org/kiwix/kiwixmobile/core/error/ErrorActivity.java
@@ -37,9 +37,9 @@ import org.kiwix.kiwixmobile.core.CoreApp;
 import org.kiwix.kiwixmobile.core.R;
 import org.kiwix.kiwixmobile.core.R2;
 import org.kiwix.kiwixmobile.core.base.BaseActivity;
+import org.kiwix.kiwixmobile.core.dao.NewBookDao;
 import org.kiwix.kiwixmobile.core.di.components.CoreComponent;
 import org.kiwix.kiwixmobile.core.entity.LibraryNetworkEntity;
-import org.kiwix.kiwixmobile.core.dao.NewBookDao;
 import org.kiwix.kiwixmobile.core.reader.ZimReaderContainer;
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem.BookOnDisk;
 
@@ -82,9 +82,8 @@ public class ErrorActivity extends BaseActivity {
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
-    injection(CoreApp.getCoreComponent());
-
     super.onCreate(savedInstanceState);
+    injection(CoreApp.getCoreComponent());
     setContentView(R.layout.activity_kiwix_error);
     Intent callingIntent = getIntent();
