+++ core/src/main/java/org/kiwix/kiwixmobile/core/history/HistoryModule.java
@@ -20,11 +20,10 @@ package org.kiwix.kiwixmobile.core.history;
 
 import dagger.Module;
 import dagger.Provides;
-import org.kiwix.kiwixmobile.core.di.ActivityScope;
 
 @Module
 public class HistoryModule {
-  //@ActivityScope
+
   @Provides
   HistoryContract.Presenter provideHistoryPresenter(HistoryPresenter presenter) {
     return presenter;
