+++ core/src/main/java/org/kiwix/kiwixmobile/core/history/HistoryPresenter.java
@@ -29,11 +29,9 @@ import java.util.List;
 import javax.inject.Inject;
 import org.kiwix.kiwixmobile.core.base.BasePresenter;
 import org.kiwix.kiwixmobile.core.data.DataSource;
-import org.kiwix.kiwixmobile.core.di.ActivityScope;
 import org.kiwix.kiwixmobile.core.di.qualifiers.Computation;
 import org.kiwix.kiwixmobile.core.di.qualifiers.MainThread;
 
-@ActivityScope
 class HistoryPresenter extends BasePresenter<HistoryContract.View>
   implements HistoryContract.Presenter {
