+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/ZimHostActivity.java
@@ -18,7 +18,6 @@
 
 package org.kiwix.kiwixmobile.core.webserver;
 
-import android.app.Activity;
 import android.app.ProgressDialog;
 import android.content.ComponentName;
 import android.content.Context;
@@ -45,7 +44,6 @@ import org.kiwix.kiwixmobile.core.CoreApp;
 import org.kiwix.kiwixmobile.core.R;
 import org.kiwix.kiwixmobile.core.R2;
 import org.kiwix.kiwixmobile.core.base.BaseActivity;
-import org.kiwix.kiwixmobile.core.di.components.ActivityComponent;
 import org.kiwix.kiwixmobile.core.di.components.CoreComponent;
 import org.kiwix.kiwixmobile.core.utils.AlertDialogShower;
 import org.kiwix.kiwixmobile.core.utils.KiwixDialog;
@@ -89,10 +87,9 @@ public class ZimHostActivity extends BaseActivity implements
 
   @Override
   protected void onCreate(@Nullable Bundle savedInstanceState) {
-    CoreApp.getCoreComponent().activityComponent().activity(this).build().inject(this);
     super.onCreate(savedInstanceState);
-    setContentView(R.layout.activity_zim_host);
     injection(CoreApp.getCoreComponent());
+    setContentView(R.layout.activity_zim_host);
 
     setUpToolbar();
