+++ core/src/main/java/org/kiwix/kiwixmobile/core/webserver/ZimHostPresenter.java
@@ -25,10 +25,8 @@ import java.util.List;
 import javax.inject.Inject;
 import org.kiwix.kiwixmobile.core.base.BasePresenter;
 import org.kiwix.kiwixmobile.core.data.DataSource;
-import org.kiwix.kiwixmobile.core.di.ActivityScope;
 import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem;
 
-@ActivityScope
 class ZimHostPresenter extends BasePresenter<ZimHostContract.View>
   implements ZimHostContract.Presenter {
